requirements.txt
fastapi
uvicorn[standard]
pydantic
retell-sdk
python-dotenv
openai
google-genai